clear & beginning of doxygen
