--- source/compression.h
@@ -1,8 +1,0 @@
-#ifndef HUFFMAN_PROJECT_COMPRESSION_H
-#define HUFFMAN_PROJECT_COMPRESSION_H
-
-#include "files/HT/HuffmanTree.h"
-
-int b_compression(char *target);
-
-#endif //HUFFMAN_PROJECT_COMPRESSION_H
