+++ source/files/UI/userInterface.h
@@ -14,13 +14,39 @@ typedef struct _MenuComponent
     int returnValue; /**< value return if component triggered */
 }MenuComponent;
 
-void HNC_help();
+/**
+ * custom_color : function that display given text with the given color
+ * @param str : text we want to display
+ * @param color : wanted color (string or string with at [0] a number)
+ */
 void custom_color(char* str, char* color);
+
+/**
+ *
+ * @param size
+ * @return
+ */
 char *cin(int size);
+
+/**
+ *
+ * @param fp
+ * @param size
+ * @return
+ */
 char *getLine(FILE *fp, int size);
+
+/**
+ * menu : bind a menu with the given component
+ * @param size : number of component
+ * @param ... : component
+ * @return the return value of the selected component (if no component selected return -1)
+ */
 int menu(int size, ...);
-int marin_pierre_babin();
-void title();
+
+/**
+ *
+ */
 void shell_tdz();
 
 #endif //HUFFMAN_PROJECT_USERINTERFACE_H
