+++ source/files/compression/compression.c
@@ -2,15 +2,107 @@
 #include <malloc.h>
 #include <string.h>
 #include "compression.h"
-#include "files/fstream/functions.h"
-#include "files/dico/dictionnary.h"
-#include "decompression.h"
-#include "files/HT/HuffmanTree.h"
+#include "../fstream/functions.h"
+#include "../dico/dictionnary.h"
+#include "../SLL/Element.h"
 
-/**
- * This function compresses a given file
- * @param target is the file we want to compress
- */
+int uncompressChar(char* target, char *tree, char *outputPath)
+{
+    FILE* treeFile = fopen(tree, "r");
+    Tree* arbre = read_tree(treeFile);
+
+    FILE* binaries = NULL;
+    binaries = fopen(target, "r");
+
+    if(!binaries)
+        return 1;
+
+    FILE* output = NULL;
+    output = fopen(outputPath, "w");
+
+    char ch = '\0';
+    Tree* searcher = arbre;
+
+    while ((ch = fgetc(binaries)) != EOF)
+    {
+        if (!hasSons(searcher))
+        {
+            fputc(searcher->data, output);
+            searcher = arbre;
+        }
+
+        if(ch == '0'){
+            searcher = searcher->left;
+        }else{
+            searcher = searcher->right;
+        }
+    }
+
+    if (!hasSons(searcher))
+    {
+        fputc(searcher->data, output);
+        searcher = arbre;
+    }
+
+    fclose(binaries);
+    fclose(output);
+    return 0;
+}
+
+void b_decompression(char *target, Tree *tree)
+{
+    FILE* binaries = NULL;
+    binaries = fopen(target, "rb");
+
+    if(!binaries)
+        return;
+
+    FILE* output = NULL;
+    output = fopen("output.txt", "w");
+
+    char ch = '\0';
+    Tree* searcher = tree;
+
+    // overbytes size
+    fread(&ch, sizeof(char), 1, binaries);
+    int bsize = (int)ch;
+    ch = '\0';
+
+    //get file size
+    int f_size = get_file_size(binaries);
+
+    // iterator
+    int i = 0;
+
+    while (ch != EOF && i < (f_size - bsize))
+    {
+        fread(&ch, sizeof(char), 1, binaries);
+        i++;
+
+        if (!hasSons(searcher))
+        {
+            //printf("%c", searcher->data);
+            fputc(searcher->data, output);
+            searcher = tree;
+        }
+
+        if(ch){
+            searcher = searcher->right;
+        }else{
+            searcher = searcher->left;
+        }
+    }
+
+    if (!hasSons(searcher))
+    {
+        fputc(searcher->data, output);
+        searcher = tree;
+    }
+
+    fclose(binaries);
+    fclose(output);
+}
+
 int b_compression(char *target)
 {
     // --- file read ---
@@ -132,9 +224,6 @@ int b_compression(char *target)
 
     // close file
     fclose(output);
-
-    // temporary
-    saveDictionary(dictionary);
 
     // free variables
     free(transcription);
