+++ source/files/fstream/functions.h
@@ -7,6 +7,11 @@
 
 #include <stdio.h>
 
+/**
+ *
+ * @param target
+ * @return
+ */
 int get_file_size(FILE* target);
 
 #endif //HUFFMAN_NAIVE_C11_FUNCTIONS_H
